Finished making the bullet mechanic

+++ Mini-project/core/src/dk/evalen19/miniprojectspaceshooter/programmingforinteraction/miniproject2020/Ship.java
@@ -30,12 +30,12 @@ public class Ship {
         this.positionY = y;
     }
 
-    public float getPositionX(){
+    public int getPositionX(){
 
         return this.positionX;
     }
 
-    public float getPositionY(){
+    public int getPositionY(){
 
         return this.positionY;
     }
